RDKB-33031: Fix warnings for Utopia
Reason for change: To fix the warnings in Utopia.
1.initialization makes integer from pointer without a cast [enabled by default]
2.missing initializer for field[-Wmissing-field-initializers]
3.implicit declaration of function [-Wimplicit-function-declaration]
4.assignment makes pointer from integer without a cast [enabled by default]
5.passing argument <argument_number> from incompatible pointer type [enabled by default]
6.format '%s' expects argument of type 'char *', but argument 3 has type 'char **' [-Wformat=]
7.conflicting types [enabled by default]
8.incompatible implicit declaration of built-in function[enabled by default]
9.passing argument <arg_num> of '<func_name>' makes pointer from integer without a cast [enabled by default]
10.second parameter of 'va_start' not last named argument [-Wvarargs]
11.argument to 'sizeof' in 'memset' call is the same expression as the destination; did you mean to dereference it? [-Wsizeof-pointer-memaccess]
12.variable set but not used [-Wunused-but-set-variable]
13.unused variable[-Wunused-variable]
14.unused parameter [-Wunused-parameter]
15.assignment discards 'const' qualifier from pointer target type [enabled by default]
16.redefined [enabled by default]
17.case value '9' not in enumerated type 'wifiTxRate_t' [-Wswitch]
18.zero-length gnu_printf format string [-Wformat-zero-length]
19.statement with no effect [-Wunused-value]
20.control reaches end of non-void function [-Wreturn-type]
21.parameter 'out_list' set but not used [-Wunused-but-set-parameter]
22.passing argument <arg_num> of <func_name> discards 'const' qualifier from pointer target type [enabled by default]
23.too many arguments for format [-Wformat-extra-args]
24.call to 'system' declared with attribute warning: please replace system() with v_secure_system() [enabled by default]
25.initialization from incompatible pointer type [enabled by default]
26.multi-line comment [-Wcomment]
27.'return' with a value, in function returning void [enabled by default]
28.suggest parentheses around assignment used as truth value [-Wparentheses]
29.extra tokens at end of #ifndef directive [enabled by default]
30.function returns address of local variable [-Wreturn-local-addr]
31.argument to 'sizeof' in 'memset' call is the same pointer type 'struct trigger_info_t *' as the destination; expected 'trigger_info_t' or an explicit length [-Wsizeof-pointer-memaccess]
32.'static' is not at beginning of declaration [-Wold-style-declaration]
33.format '%d' expects a matching 'int' argument [-Wformat=]
34.comparison between pointer and integer [enabled by default]
35.array subscript has type 'char' [-Wchar-subscripts]
36.unknown escape sequence: '\.' [enabled by default]
37.the address of 'router' will always evaluate as 'true' [-Waddress]
38.too many arguments for format [-Wformat-extra-args]
Test Procedure: Build should be successful and basic functionality of Utopia should be verified.
Risks: High

(cherry picked from commit 76b58ce8e94998a741f1a019e47678131c905fdc)

+++ source/utapi/lib/tr181_wrpr.c
@@ -242,7 +242,7 @@ int Utopia_Get_TR181_Device_MoCA_Interface_i_Dyn(Obj_Device_MoCA_Interface_i_dyn
     ptr = head = NULL;
     
     memset(buf, 0, BUF_SZ);
-    sprintf (buf, "cat %s|grep -A %d 'NodeID:%d' > %s", MOCA_ASSOC_DEV, INST_SIZE, deviceMocaIntfDyn->NodeID, MOCA_MAC_NODE);
+    sprintf (buf, "cat %s|grep -A %d 'NodeID:%d' > %s", MOCA_ASSOC_DEV, INST_SIZE, (int)deviceMocaIntfDyn->NodeID, MOCA_MAC_NODE);
     system(buf);
     
     retVal = file_parse(MOCA_MAC_NODE, &head);
